+++ ididi/__init__.py
@@ -18,17 +18,19 @@ VERSION = "1.1.2"
 
 __version__ = VERSION
 
+from ._itypes import EntryConfig as EntryConfig
 from ._itypes import GraphConfig as GraphConfig
+from ._itypes import IEntryConfig as IEntryConfig
 from ._itypes import INode as INode
 from ._itypes import INodeConfig as INodeConfig
 from ._itypes import NodeConfig as NodeConfig
-from ._itypes import IEntryConfig as IEntryConfig
-from ._itypes import EntryConfig as EntryConfig
 from .api import entry as entry
 from .api import resolve as resolve
+from .graph import AsyncScope as AsyncScope
 from .graph import DependencyGraph as DependencyGraph
+from .graph import SyncScope as SyncScope
 from .node import DependentNode as DependentNode
-from .node import inject as inject
+from .node import use as use
 from .utils.typing_utils import AsyncResource as AsyncResource
 from .utils.typing_utils import Resource as Resource